Update RetrofitClientInstance.kt BASE_URL
I changed the type of BASE_URL to include the const keyword. While not necessary to compile, const is used to declare a read-only property of a class. Since we are using retrofit to import mock JSON data right now, we do not need this to be a read/write variable. Const ensures that we cannot accidentally modify the JSON data.