Update ArticleUnitTest.kt
Changed equals() statements in the thenResultsShouldContainRihanna function to use ==. Both check structural equality but I think == makes the statement a little more readable by removing excess parentheses.

Added parentheses around individual statements in the if() statement for syntax and separation. May be redundant in reference to the above listed change.